Revert rename of tracking_graph
