+++ src/cell_tracking_metrics/__init__.py
@@ -9,7 +9,7 @@ except PackageNotFoundError:
 __author__ = "Track Gals"
 __email__ = "draga.doncilapop1@monash.edu"
 
-from ._tracking_graph import TrackingGraph
 from .tracking_data import TrackingData
+from .tracking_graph import TrackingGraph
 
 __all__ = ["TrackingData", "TrackingGraph"]
